+++ classes/App.class.php
@@ -159,18 +159,6 @@ class App
             'exp_msg'       => $M->isNew() ? '' :
                 sprintf($LANG_MEMBERSHIP['you_expire'], $M->getPlanID(), $M->getExpires()),
         ) );
-        if ($_CONF_MEMBERSHIP['update_maillist']) {
-            $status = PLG_invokeService('mailchimp', 'issubscribed',
-                array(
-                    'uid'=>$this->uid,
-                ),
-                $output,
-                $svc_msg
-            );
-            if ($status == PLG_RET_OK && $output) {
-                $T->set_var('update_maillist', 'true');
-            }
-        }
 
         // Add the checkbox to accept terms or waivers, if so configured
         if ($_CONF_MEMBERSHIP['terms_url'] != '') {
@@ -280,9 +268,11 @@ class App
                 // Save and log the terms and conditions acceptance.
                 // Subscribe the user to the default mailing list
                 // if selected
-                if (isset($_POST['mailchimp_subscribe']) &&
-                    $_POST['mailchimp_subscribe'] == 1) {
-                    PLG_invokeService('mailchimp', 'subscribe',
+                if (
+                    isset($_POST['maillist_subscribe']) &&
+                    $_POST['maillist_subscribe'] == 1
+                ) {
+                    PLG_invokeService('mailer', 'subscribe',
                         array(
                             'uid'=> $uid,
                         ),