+++ services.inc.php
@@ -192,18 +192,22 @@ function service_profilefilter_membership($args, &$output, &$svc_msg)
     global $LANG_MEMBERSHIP, $_CONF_MEMBERSHIP;
 
     // Non-managers don't get access to view other members' expiration
-    if (!MEMBERSHIP_isManager()) return PLG_RET_NOACCESS;
+    if (!MEMBERSHIP_isManager()) return PLG_RET_PERMISSION_DENIED;
 
     $opts = array(
-        Status::ENABLED => $LANG_MEMBERSHIP['current'],
+        Status::ACTIVE => $LANG_MEMBERSHIP['current'],
         Status::ARREARS => $LANG_MEMBERSHIP['arrears'],
-        //Status::EXPIRED => $LANG_MEMBERSHIP['expired'],
+        Status::EXPIRED => $LANG_MEMBERSHIP['expired'],
     );
+
     $output = array();
     // If posted variables are recieved, use them. Otherwise, use GET but only
     // if POST is empty. Otherwise the user may have just unchecked all the
     // options
-    if (isset($args['post']['mem_exp_status_flag'])) {
+    if (
+        isset($args['post']['mem_exp_status_flag']) &&
+        isset($args['post']['mem_exp_status'])
+    ) {
         $exp_stat = $args['post']['mem_exp_status'];
     } elseif (empty($args['post']) && isset($args['get']['mem_exp_status'])) {
         $exp_stat = explode(',', $args['get']['mem_exp_status']);
@@ -218,7 +222,7 @@ function service_profilefilter_membership($args, &$output, &$svc_msg)
         // Use the default setting if no other options received
         $exp_stat = array();
         if ($_CONF_MEMBERSHIP['prflist_current'] == 1)
-            $exp_stat[] = Status::ENABLED;
+            $exp_stat[] = Status::ACTIVE;
         if ($_CONF_MEMBERSHIP['prflist_arrears'] == 1)
             $exp_stat[] = Status::ARREARS;
         //if ($_CONF_MEMBERSHIP['prflist_expired'] == 1)
@@ -274,7 +278,10 @@ function service_profilefields_membership($args, &$output, &$svc_msg)
     }
 
     // Include the expiration status, if requested
-    if (isset($args['post']['mem_exp_status_flag'])) {
+    if (
+        isset($args['post']['mem_exp_status_flag']) &&
+        isset($args['post']['mem_exp_status'])
+    ) {
         $exp_stat = $args['post']['mem_exp_status'];
     } elseif (empty($args['post']) && isset($args['get']['mem_exp_status'])) {
         $exp_stat = explode(',', $args['get']['mem_exp_status']);
@@ -282,7 +289,7 @@ function service_profilefields_membership($args, &$output, &$svc_msg)
         $exp_stat = array();
         if (is_int($args['incl_exp_stat'])) {
             foreach (array(
-                Status::ENABLED,
+                Status::ACTIVE,
                 Status::ARREARS,
                 Status::EXPIRED,
             ) as $key) {
@@ -291,8 +298,8 @@ function service_profilefields_membership($args, &$output, &$svc_msg)
                 }
             }
         } elseif (is_array($args['incl_exp_stat'])) {
-            if ($args['incl_exp_stat'] && Status::ENABLED) {
-                $exp_stat[] = Status::ENABLED;
+            if ($args['incl_exp_stat'] && Status::ACTIVE) {
+                $exp_stat[] = Status::ACTIVE;
             }
             if ($args['incl_exp_stat'] && Status::ARREARS) {
                 $exp_stat[] = Status::ARREARS;
@@ -311,7 +318,7 @@ function service_profilefields_membership($args, &$output, &$svc_msg)
             // Only create sql if filtering on some expiration status
             $grace = (int)$_CONF_MEMBERSHIP['grace_days'];
             $exp_arr = array();
-            if ($incl_exp_stat & Status::ENABLED == Status::ENABLED) {
+            if ($incl_exp_stat & Status::ACTIVE == Status::ACTIVE) {
                 $exp_arr[] = "$members.mem_expires >= '" . Dates::Today() . "'";
             }
             if ($incl_exp_stat & Status::ARREARS) {
@@ -532,7 +539,7 @@ function service_ismember_membership($args, &$output, &$svc_msg)
 function service_mailingSegment_membership($args, &$output, &$svc_msg)
 {
     global $_TABLES;
-var_dump($args);die;
+
     // Get the current statuses
     //$statuses = MEMBERSHIP_memberstatuses();
 
@@ -541,18 +548,24 @@ var_dump($args);die;
     $uid = 0;
 
     if (isset($args['email']) && !empty($args['email'])) {
-        $uid = (int)DB_getItem($_TABLES['users'], 'uid',
-                "email = '" . DB_escapeString($args['email']) . "'");
+        $uid = (int)DB_getItem(
+            $_TABLES['users'],
+            'uid',
+            "email = '" . DB_escapeString($args['email']) . "'"
+        );
     } elseif (isset($args['uid']) && $args['uid'] > 1) {
         $uid = (int)$args['uid'];
     }
 
     if ($uid > 0) {
+        $output = plugin_getiteminfo_membership('membership:' . $uid, 'id,merge_fields');
+        /*
         $myargs = array('uid' => $uid);
         $code = service_status_membership($myargs, $myout, $svc_msg);
         if ($code == PLG_RET_OK && isset($statuses[$myout['status']])) {
             $output = Status::getSegment($myout['status']);
         }
+         */
     }
     return PLG_RET_OK;
 }